+++ kol/formats/mjcf.py
@@ -5,6 +5,8 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Literal
 
+from kol.formats.common import save_xml
+
 
 @dataclass
 class Compiler:
@@ -48,23 +50,4 @@ class Robot:
 
     def save(self, path: str | Path) -> None:
         tree = ET.ElementTree(self.to_xml())
-        root = tree.getroot()
-
-        def indent(elem: ET.Element, level: int = 0) -> ET.Element:
-            i = "\n" + level * "  "
-            if len(elem):
-                if not elem.text or not elem.text.strip():
-                    elem.text = i + "  "
-                if not elem.tail or not elem.tail.strip():
-                    elem.tail = i
-                for elem in elem:
-                    indent(elem, level + 1)
-                if not elem.tail or not elem.tail.strip():
-                    elem.tail = i
-            else:  # noqa: PLR5501
-                if level and (not elem.tail or not elem.tail.strip()):
-                    elem.tail = i
-            return elem
-
-        indent(root)
-        tree.write(path, encoding="utf-8", xml_declaration=True, method="xml")
+        save_xml(path, tree)
